WIP: conversion to native python types

+++ python-tests/test_rle.py
@@ -4,13 +4,13 @@ import unittest
 import numpy as np
 
 def test_rle():
-    data = np.array([0, 0, 0, 1, 1, 1, 1, 0, 0, 0])
+    data = np.array([0, 0, 0, 1, 1, 1, 1, 0, 0, 0], dtype=np.uint64)
     output = up.rle_encode(data)
     assert output == [3, 0, 4, 1, 3, 0]
     
 
 def test_rle_multi():
-    data = np.array([[0, 0, 0, 1, 1, 1, 1, 0, 0, 0],[0, 0, 0, 1, 1, 1, 1, 0, 0, 0]])
+    data = np.array([[0, 0, 0, 1, 1, 1, 1, 0, 0, 0],[0, 0, 0, 1, 1, 1, 1, 0, 0, 0]], dtype=np.uint64)
     output = up.rle_encode(data)
     assert output == [3, 0, 4, 1, 6, 0, 4, 1, 3, 0]
